almost done

+++ Calc/Goods.cs
@@ -9,13 +9,13 @@ namespace Calc
 {
     public class Goods
     {
-        public int Weight { get; set; }
-        public int Length { get; set; }
-        public int Height { get; set; }
-        public int Width  { get; set; }
+        public double Weight { get; set; }
+        public double Length { get; set; }
+        public double Height { get; set; }
+        public double Width  { get; set; }
         public string Name { get; set; }
 
-        public Goods(string Name, int Weight, int Length, int Height, int Width)
+        public Goods(string Name, double Weight, double Length, double Height, double Width)
         {
             this.Weight = Weight;
             this.Length = Length;
